chore: move remote tutorials
